:recycle: Refactor FetchUtils to separate functions
